requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.3.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2020.12.5
cffi==1.14.5
channels==3.0.4
channels-redis==3.2.0
chardet==4.0.0
chromedriver-py==100.0.4896.20
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
configparser==5.2.0
constantly==15.1.0
crayons==0.4.0
cryptography==3.4.7
daphne==3.0.2
Deprecated==1.2.13
distro==1.6.0
Django==3.2.7
django-channels==0.7.0
django-filter==21.1
django-oauth-toolkit==1.5.0
django-rest-framework==0.1.0
django-static-jquery==2.1.4
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-filters==0.11.1
et-xmlfile==1.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.6.1
incremental==21.3.0
jwcrypto==1.0
kombu==5.1.0
msgpack==1.0.2
mysqlclient==2.0.3
numpy==1.19.5
oauthlib==3.1.1
openpyxl==3.0.7
pandas==1.1.5
pdfminer==20191125
prompt-toolkit==3.0.19
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.11.0
pyOpenSSL==20.0.1
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==303
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tabula-py==2.3.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.5.3
websocket-client==1.1.0
wget==3.2
wrapt==1.13.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.5.0
zope.interface==5.4.0
